+++ floor_generator/lib/misc/extension/list_equality_extension.dart
@@ -1,0 +1,5 @@
+import 'package:collection/collection.dart';
+
+extension ListEqualityExtension<T> on List<T> {
+  bool equals(List<T> other) => ListEquality<T>().equals(this, other);
+}