+++ floor_generator/test/value_object/entity_test.dart
@@ -20,6 +20,7 @@ void main() {
     'field1ColumnName',
     false,
     SqlType.integer,
+    null,
   );
   final nullableField = Field(
     mockFieldElement,
@@ -27,6 +28,7 @@ void main() {
     'field2ColumnName',
     true,
     SqlType.text,
+    null,
   );
   final allFields = [field, nullableField];
 
@@ -51,6 +53,7 @@ void main() {
         [],
         false,
         '',
+        '',
       );
 
       final actual = entity.getCreateTableStatement();
@@ -72,6 +75,7 @@ void main() {
         [],
         [],
         false,
+        '',
         '',
       );
 
@@ -95,6 +99,7 @@ void main() {
         [],
         [],
         false,
+        '',
         '',
       );
 
@@ -128,6 +133,7 @@ void main() {
         [],
         false,
         '',
+        '',
       );
 
       final actual = entity.getCreateTableStatement();
@@ -155,6 +161,7 @@ void main() {
       [],
       true,
       '',
+      '',
     );
 
     final actual = entity.getCreateTableStatement();
@@ -165,69 +172,5 @@ void main() {
         'PRIMARY KEY (`${field.columnName}`)'
         ') WITHOUT ROWID';
     expect(actual, equals(expected));
-  });
-
-  group('Value mapping', () {
-    final primaryKey = PrimaryKey([nullableField], true);
-    final entity = Entity(
-      mockClassElement,
-      'entityName',
-      [nullableField],
-      primaryKey,
-      [],
-      [],
-      false,
-      '',
-    );
-    const fieldElementDisplayName = 'foo';
-
-    setUp(() {
-      when(mockFieldElement.displayName).thenReturn(fieldElementDisplayName);
-      when(mockFieldElement.type).thenReturn(mockDartType);
-    });
-
-    test('Get value mapping', () {
-      when(mockDartType.isDartCoreBool).thenReturn(false);
-
-      final actual = entity.getValueMapping();
-
-      final expected = '<String, dynamic>{'
-          "'${nullableField.columnName}': item.$fieldElementDisplayName"
-          '}';
-      expect(actual, equals(expected));
-    });
-
-    test('Get nullable boolean value mapping', () {
-      when(mockDartType.isDartCoreBool).thenReturn(true);
-
-      final actual = entity.getValueMapping();
-
-      final expected = '<String, dynamic>{'
-          "'${nullableField.columnName}': item.$fieldElementDisplayName == null ? null : (item.$fieldElementDisplayName ? 1 : 0)"
-          '}';
-      expect(actual, equals(expected));
-    });
-
-    test('Get non-nullable boolean value mapping', () {
-      final entity = Entity(
-        mockClassElement,
-        'entityName',
-        [nullableField, field],
-        primaryKey,
-        [],
-        [],
-        false,
-        '',
-      );
-      when(mockDartType.isDartCoreBool).thenReturn(true);
-
-      final actual = entity.getValueMapping();
-
-      final expected = '<String, dynamic>{'
-          "'${nullableField.columnName}': item.$fieldElementDisplayName == null ? null : (item.$fieldElementDisplayName ? 1 : 0),"
-          " '${field.columnName}': item.$fieldElementDisplayName ? 1 : 0"
-          '}';
-      expect(actual, equals(expected));
-    });
   });
 }